% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{compare_type}
\alias{compare_type}
\title{Draw the posterior distribution at the group level}
\usage{
compare_type(fit, xlim = NULL, ylim = NULL)
}
\arguments{
\item{fit}{A stanfit object}

\item{xlim}{Range of the x-axis}

\item{ylim}{Range of the y-axis}
}
\value{
An object of class ggExtraPlot
}
\description{
Draw the posterior distribution at the group level
}
