% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_data.R
\name{correct_lambda}
\alias{correct_lambda}
\title{Add a column named lambda to the data.frame}
\usage{
correct_lambda(dat, change_lambda)
}
\arguments{
\item{dat}{The data.frame to where we will add the lambda column}

\item{change_lambda}{The data.frame explaining where lambda should not
be equal to 1}
}
\value{
The input dat completed with a new column named lambda.
}
\description{
Add a column named lambda to the data.frame
}
