% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{envelop}
\alias{envelop}
\title{Draw posterior Gompertz curve envelop}
\usage{
envelop(
  fit,
  data,
  id,
  t = seq(0, 11, length.out = 101),
  ci_level = 0.5,
  color = "#4B8EC7"
)
}
\arguments{
\item{fit}{A stanfint object}

\item{data}{The dataset}

\item{id}{A vector of length 1, containing the id of the individual to plot}

\item{t}{The time grid}

\item{ci_level}{Posterior probability of the ribbon of darkest color}

\item{color}{Color of the ribbon}
}
\value{
A ggplot object
}
\description{
Draw posterior Gompertz curve envelop
}
