% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scripts_stan.R
\name{init_fun}
\alias{init_fun}
\title{Propose initial values for the MCMC}
\usage{
init_fun(par_LS, data_stan, z0.global = TRUE, ...)
}
\arguments{
\item{par_LS}{the least square estimates}

\item{data_stan}{the list, as outputed by set_data}

\item{z0.global}{If FALSE, assume a group hierachical prior on z0}
}
\value{
A list to be used by stan
}
\description{
Propose initial values for the MCMC
}
