% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/least_square.R
\name{least_square}
\alias{least_square}
\title{Fit Gompertz curves with a least square estimate}
\usage{
least_square(data, lambda = NULL, parinit = c(20, 10, 0.5))
}
\arguments{
\item{data}{The data.set, in long format}

\item{lambda}{A vector to change the weight of a few time points}

\item{parinit}{The initial parameter value}
}
\value{
The least square estimates for each individual in the dataset
}
\description{
Fit Gompertz curves with a least square estimate
}
