% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_data.R
\name{long_log_format}
\alias{long_log_format}
\title{Transform dataset to a long format}
\usage{
long_log_format(dat)
}
\arguments{
\item{dat}{A data.frame}
}
\value{
A data.frame in long format.
}
\description{
Transform dataset to a long format
}
\details{
Assume the 1st column of the input is a type, and all other columns
are observed values at different times. Names of thoses columns should be
the time, prefixed by "T"
}
