% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot_LS}
\alias{plot_LS}
\title{Plot the dataset and the fitted Gompertz curves}
\usage{
plot_LS(dat, par_LS, n = 101)
}
\arguments{
\item{dat}{The dataset}

\item{par_LS}{The least square estimates}

\item{n}{Number of points to draw the Gompertz curves}
}
\value{
A ggplot object
}
\description{
Plot the dataset and the fitted Gompertz curves
}
