% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scripts_stan.R
\name{set_data}
\alias{set_data}
\title{Prepare a list of data for Stan}
\usage{
set_data(dat, z0.global = TRUE, groups.explain = TRUE)
}
\arguments{
\item{dat}{The dataset}

\item{z0.global}{If FALSE, assume a group hierachical prior on z0}

\item{groups.explain}{If TRUE, write how the type were coded by numbers.}
}
\value{
A list ready to be passed to stan
}
\description{
Prepare a list of data for Stan
}
